multiplayer role usage in favour of defines
