+++ Experimental/DedicatedGameServer/ProjectSettings/MultiplayerManager.asset
@@ -6,4 +6,77 @@ MultiplayerManager:
   m_EnableMultiplayerRoles: 1
   m_StrippingTypes:
     1: []
-    2: []
+    2:
+    - UnityEngine.UI:UnityEngine.UI:PositionAsUV1
+    - UnityEngine.UI:UnityEngine.EventSystems:TouchInputModule
+    - UnityEngine.UI:UnityEngine.UI:Toggle
+    - UnityEngine.UI:UnityEngine.UI:ContentSizeFitter
+    - UnityEngine.UI:UnityEngine.UI:CanvasScaler
+    - UnityEngine.UI:UnityEngine.UIElements:PanelEventHandler
+    - UnityEngine.CoreModule:UnityEngine:Light
+    - UnityEngine.UI:UnityEngine.EventSystems:PointerInputModule
+    - UnityEngine.UI:UnityEngine.UI:LayoutGroup
+    - Unity.RenderPipelines.Universal.Runtime:UnityEngine.Rendering.Universal:UniversalAdditionalCameraData
+    - UnityEngine.UI:UnityEngine.UI:ToggleGroup
+    - UnityEngine.UIElementsModule:UnityEngine.UIElements:InputWrapper
+    - UnityEngine.UI:UnityEngine.UI:Slider
+    - UnityEngine.UI:UnityEngine.UI:Shadow
+    - UnityEngine.AudioModule:UnityEngine:AudioLowPassFilter
+    - UnityEngine.SpriteShapeModule:UnityEngine.U2D:SpriteShapeRenderer
+    - UnityEngine.AudioModule:UnityEngine:AudioSource
+    - UnityEngine.AudioModule:UnityEngine:AudioHighPassFilter
+    - UnityEngine.UI:UnityEngine.UI:HorizontalLayoutGroup
+    - UnityEngine.UI:UnityEngine.UI:GraphicRaycaster
+    - UnityEngine.UI:UnityEngine.UI:Dropdown
+    - UnityEngine.UI:UnityEngine.UI:BaseMeshEffect
+    - UnityEngine.UI:UnityEngine.UIElements:PanelRaycaster
+    - UnityEngine.UIElementsModule:UnityEngine.UIElements:UIDocument
+    - UnityEngine.UIElementsModule:UnityEngine.UIElements.InputSystem:InputSystemEventSystem
+    - UnityEngine.ParticleSystemModule:UnityEngine:ParticleSystemRenderer
+    - UnityEngine.CoreModule:UnityEngine:MeshRenderer
+    - UnityEngine.CoreModule:UnityEngine:Camera
+    - UnityEngine.SpriteMaskModule:UnityEngine:SpriteMask
+    - UnityEngine.CoreModule:UnityEngine:SpriteRenderer
+    - UnityEngine.TilemapModule:UnityEngine.Tilemaps:TilemapRenderer
+    - UnityEngine.UI:UnityEngine.UI:AspectRatioFitter
+    - UnityEngine.UIElementsModule:UnityEngine.UIElements:UIRenderer
+    - UnityEngine.AudioModule:UnityEngine:AudioEchoFilter
+    - UnityEngine.UI:UnityEngine.EventSystems:PhysicsRaycaster
+    - UnityEngine.UI:UnityEngine.EventSystems:BaseInputModule
+    - UnityEngine.UI:UnityEngine.UI:Dropdown/DropdownItem
+    - UnityEngine.UI:UnityEngine.UI:HorizontalOrVerticalLayoutGroup
+    - UnityEngine.UI:UnityEngine.EventSystems:BaseInput
+    - UnityEngine.UI:UnityEngine.UI:InputField
+    - UnityEngine.UI:UnityEngine.EventSystems:EventSystem
+    - UnityEngine.UI:UnityEngine.UI:RawImage
+    - UnityEngine.UI:UnityEngine.EventSystems:StandaloneInputModule
+    - UnityEngine.UI:UnityEngine.UI:Selectable
+    - UnityEngine.UI:UnityEngine.EventSystems:UIBehaviour
+    - UnityEngine.UI:UnityEngine.EventSystems:BaseRaycaster
+    - UnityEngine.UI:UnityEngine.UI:Image
+    - UnityEngine.UI:UnityEngine.UI:MaskableGraphic
+    - UnityEngine.AudioModule:UnityEngine:AudioBehaviour
+    - UnityEngine.UI:UnityEngine.UI:Scrollbar
+    - UnityEngine.UI:UnityEngine.UI:Text
+    - UnityEngine.UI:UnityEngine.UI:GridLayoutGroup
+    - UnityEngine.UI:UnityEngine.UI:VerticalLayoutGroup
+    - UnityEngine.UI:UnityEngine.UI:Mask
+    - UnityEngine.AudioModule:UnityEngine:AudioListener
+    - UnityEngine.CoreModule:UnityEngine:BillboardRenderer
+    - UnityEngine.UI:UnityEngine.UI:Button
+    - UnityEngine.UI:UnityEngine.UI:LayoutElement
+    - UnityEngine.UI:UnityEngine.EventSystems:Physics2DRaycaster
+    - UnityEngine.AudioModule:UnityEngine:AudioDistortionFilter
+    - UnityEngine.CoreModule:UnityEngine:SkinnedMeshRenderer
+    - UnityEngine.UI:UnityEngine.UI:Graphic
+    - UnityEngine.UI:UnityEngine.UI:RectMask2D
+    - UnityEngine.AudioModule:UnityEngine:AudioReverbZone
+    - UnityEngine.CoreModule:UnityEngine:TrailRenderer
+    - UnityEngine.AudioModule:UnityEngine:AudioReverbFilter
+    - UnityEngine.AudioModule:UnityEngine:AudioChorusFilter
+    - UnityEngine.UI:UnityEngine.EventSystems:EventTrigger
+    - UnityEngine.VFXModule:UnityEngine.VFX:VFXRenderer
+    - UnityEngine.CoreModule:UnityEngine:LineRenderer
+    - UnityEngine.UI:UnityEngine.UI:Outline
+    - Unity.RenderPipelines.Universal.Runtime:UnityEngine.Rendering.Universal:UniversalAdditionalLightData
+    - UnityEngine.UI:UnityEngine.UI:ScrollRect
